requirements.txt
fastapi
pytest
uvicorn
torch
diffusers
transformers
pillow
python-dotenv
git+https://github.com/openai/CLIP.git
